==> 02_model_construction.ipynb <==
# Model construction

# import required libraries
import numpy as np
import pandas as pd
import pathlib
from sklearn.metrics import classification_report, f1_score
import matplotlib.pyplot as plt
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from assets.functions import model_quality_assurance_preprocessing

# load preprocessed dataset
raw_data_preprocessed = pd.read_pickle(pathlib.Path(r'data\\processed_data\\raw_data_preprocessed.pkl'))
# Get the list of unique indices
indices = raw_data_preprocessed["measurement_data_compensated"].index.get_level_values(0).unique()
# Use groupby to split the DataFrame
grouped = raw_data_preprocessed["measurement_data_compensated"].groupby(level=0)
# get fcs data as a list of dataframes of each sample
fcs_data = [grouped.get_group(index).reset_index(drop=True) for index in indices]

# same for labels
indices = raw_data_preprocessed["reported_labels"].index.get_level_values(0).unique()
# Use groupby to split the DataFrame
grouped = raw_data_preprocessed["reported_labels"].groupby(level=0)
# get fcs data as a list of dataframes of each sample
labelset = [grouped.get_group(index).reset_index(drop=True) for index in indices]

# train test split and save respective indices

files_used_for_training_of_RFC = '2000 events of each file, from 0 to 50, as a single long list'

cell_subsets_of_interest = ['Lympho', 'BP', 'NKP', 'TP', 'T4P', 'T8P']

# to reduce computational costs reduce subsample the training set
train_size = 2000

x_train, x_valid, y_train, y_valid, indices_train, indices_valid = [], [], [], [], [], []
for i in range(0, 50):
    (x_train_loop, x_valid_loop, 
     y_train_loop, y_valid_loop,
    indices_train_loop, indices_valid_loop) = train_test_split(fcs_data[i], 
                                                               labelset[i][cell_subsets_of_interest],
                                                               labelset[i][cell_subsets_of_interest].index,
                                                               train_size=train_size, 
                                                               test_size=len(fcs_data[i])-train_size,
                                                               random_state=42)
    illegal_indices_train = y_train_loop.query('BP == 1 and NKP == 1').index
    illegal_indices_valid = y_valid_loop.query('BP == 1 and NKP == 1').index
    x_train.append(x_train_loop.drop(illegal_indices_train))
    y_train.append(y_train_loop.drop(illegal_indices_train))
    x_valid.append(x_valid_loop.drop(illegal_indices_valid))
    y_valid.append(y_valid_loop.drop(illegal_indices_valid))
    indices_train.append(indices_train_loop.drop(list(illegal_indices_train)))
    indices_valid.append(indices_valid_loop.drop(list(illegal_indices_valid)))

# use remaining files as testing
x_test, y_test = [], []
for i in range(50, len(fcs_data)):
    x_test_loop = fcs_data[i]
    y_test_loop = labelset[i]
    x_test.append(x_test_loop)
    y_test.append(y_test_loop)

# data preparation, please see model_development.parameter_optimization.ipynb
transformation = 'fluorescence channels transformed by \n None'
scaling = 'channels scaled by \n None'
domain_adaptation = 'domain adaptation \n no method applied'
best_rfc = pd.read_pickle(pathlib.Path(r'model_development\\grid_search_results\\rfc.pkl')).sort_values(by='score', ascending=False).iloc[0]
best_rfc

best_rfc['params']

# use best performing model and parameters, please see model_development.parameter_optimization.ipynb
rfc = RandomForestClassifier(n_estimators=120, random_state=42, n_jobs=-1)
rfc.fit(pd.concat(x_train), pd.concat(y_train))

# classify all events for subsequent model quality assurance
# predictions on the training and validation set is required for uncertainty estimation

rfc_prediction = []
for i in range(100):
    prediction_loop = rfc.predict(fcs_data[i])
    prediction_loop = pd.DataFrame(prediction_loop, 
                                   columns=['pred_Lympho','pred_BP', 'pred_NKP', 'pred_TP', 
                                            'pred_T4P', 'pred_T8P'])
    rfc_prediction.append(prediction_loop)
    print(i, 'done!', end="\r")

# calculate cell type specific f1 scores

f1_scores = []
for i, j in zip(labelset, rfc_prediction):
    f1_scores.append(pd.DataFrame(
        f1_score(y_true=i[cell_subsets_of_interest], y_pred=j, average=None)[np.newaxis],
        columns=i[cell_subsets_of_interest].columns
        ))
f1_scores = pd.concat(f1_scores, ignore_index=True)

# save f1 scores

f1_scores.to_pickle(r'data/processed_data/f1_scores_model.pkl')

# provide the indices used for training
for i in range(50):
    indices_train.append(range(0))


# preprocess data for subsequent analyses
dataset_model_predictions =[]
for i in range(100):
    dataset_model_predictions.append(
        model_quality_assurance_preprocessing(fcs_data[i], 
                                              labelset[i],
                                              indices_train[i], 
                                              rfc_prediction[i], 
                                              i)
    )

# show the processed df
data_model = pd.concat(dataset_model_predictions)
data_model

data_model.to_pickle('data\\processed_data\\#1_aicontrol_dataset_pred.pkl')
