==> 03_model_quality_assurance_preprocessing.ipynb <==
# Prepare data for model quality assurance

import pandas as pd
import pathlib
import numpy as np
from assets.hierarchy import get_hierarchy_lookup
from assets.functions import gate_for_patient

# Load data
data_model = pd.read_pickle(
    pathlib.Path(r'data\\processed_data\\#1_aicontrol_dataset_pred.pkl')
    )

# convert multiindex into columns
data_model.reset_index(inplace=True, drop=False, col_level=1)
# drop header to easy data handling
data_model = data_model.droplevel(axis=1, level=0)

data_model

# get traning data
data_train = data_model.loc[data_model['used_for_training_Q_file'] == 1]
# Get patient IDs of events used for training -> should not be used for quality assurance
patients_train = data_train.loc[:, 'file'].unique()
# get test data
data_test = data_model[~data_model['file'].isin(patients_train)]

# save training and test datasets
data_train.to_pickle(pathlib.Path(r'data/processed_data/dataset_pred_train.pkl'))
data_test.to_pickle(pathlib.Path(r'data/processed_data/dataset_pred_test.pkl'))

# save complete dataset (reformatted)
data_model.to_pickle(pathlib.Path(r'data/processed_data/dataset_pred_complete.pkl'))

# save patient IDs of training events
pd.Series(patients_train).to_pickle(pathlib.Path(r'data/processed_data/patient_ids_train.pkl'))

## Sum events per sample

# list of celltypes considered by the model
celltypes_ddm = ['Lympho', 'BP','NKP', 'TP', 'T4P', 'T8P']

# list of column names for predicted events by the model
celltypes_pred = [f'pred_{celltype}' for celltype in celltypes_ddm]

# names of columns that will be summed
sum_cols = celltypes_ddm + celltypes_pred

event_sum_per_sample = data_model.loc[:, ['file'] + sum_cols].groupby('file').sum()

event_sum_per_sample.loc[:, 'total'] = data_model.loc[:, ['file'] + sum_cols].groupby('file').size()

event_sum_per_sample

event_sum_per_sample.to_pickle(pathlib.Path(r'data/processed_data/event_sum_per_sample_complete.pkl'))

celltypes_pred.append('total')
event_sum_per_sample[celltypes_pred].to_pickle(pathlib.Path(r'data/processed_data/model_summed.pkl'))

## Determine manual linear gating

# Names of the marker features
marker_names = ['FSC-A', 'SSC-A', 'APC-A', 'APC-H7-A', 'AmCyan-A', 'FITC-A', 'PE-A', 
                'PE-Cy7-A', 'Pacific Blue-A', 'PerCP-A']

# Names of the ground truth features of all cell types
gt_names_all = ['BP', 'CD4', 'CD4x38',
       'CD4xDR', 'CD4xDRx38', 'CD8', 'CD8x38', 'CD8xDR', 'CD8xDRx38', 'Lympho',
       'Manja rufen!', 'NKP', 'NKTP', 'NKx38', 'NKxDR', 'NKxDRx38', 'Q1-1',
       'Q1-2', 'Q1-3', 'Q1-5', 'Q1-6', 'Q1-7', 'Q1-9', 'Q2', 'Q3', 'Q3-1',
       'Q3-2', 'Q3-4', 'Q3-5', 'Q3-6', 'Q3-7', 'Q3-9', 'Q4-1', 'Q4-2', 'Q4-3',
       'Q4-4', 'Q4-5', 'Q4-6', 'Q4-7', 'Q4-9', 'T48NT', 'T48PT', 'T4P', 'T8P',
       'TBNK Sum', 'TP']

# Names of the ground truth features of selected cell types, i.e. cell types that are also predicted by the DDM
gt_names_selected = ['Lympho', 'T4P', 'T8P', 'BP', 'TP', 'NKP']

# Names of the features for the predicted cell types
pred_names_selected = ['pred_{}'.format(celltype) for celltype in gt_names_selected]

hierarchy_lookup = get_hierarchy_lookup()

# Names of celltypes determined in gating plots
gating_celltype_names = list(hierarchy_lookup.loc[hierarchy_lookup['marker_x'].notna()].index.values)
gating_celltype_names.remove('Lympho')
gating_celltype_names.remove('Lympho_alternative')

gating_celltype_names.remove('NKxDRx38')
gating_celltype_names.remove('CD4xDRx38')
gating_celltype_names.remove('CD8xDRx38')

# create column Not(TP)
data_model.loc[:, 'Not_TP'] = (~data_model.loc[:, 'TP'].astype(bool)).astype(np.uint8)

# determine the x and y axis position of linear gates
gating_df = []

for patient_id in data_model['file'].unique():
    gating_df.append(gate_for_patient(patient_id, data_model))
    
gating_df = pd.concat(gating_df)

gating_df

gating_df.to_pickle(pathlib.Path(r'data/processed_data/manual_gating_expert1.pkl'))
